feat: 유효성 검사 함수 세부 분기 구현 및 타입과 유효성 검사 함수 파일 분할
